Added logging, refactored for parallel structure, reworded comments
